Detect Docker Path Mapping
The plugin attempts to automatically detect if running inside a docker
container with path mounted in a different location, it will apply
the mappings to the path to report them back to Azure Pipelines using the path
from the host that has been bind mounted to the docker container. No
configuration is required it should just work as long as bind mounting is
used to the path the pytest output is written to. Also ensure the files are
written using an account the host has access to, this can be done by supplying
the user and group of the host account to the run command.